+++ pkg/order/etcdraft/node_test.go
@@ -86,7 +86,7 @@ func generateTx() *pb.Transaction {
 			Amount: 10,
 		},
 		Timestamp: time.Now().UnixNano(),
-		Nonce:     rand.Int63(),
+		Nonce:     uint64(rand.Int63()),
 	}
 	_ = tx.Sign(privKey)
 	tx.TransactionHash = tx.Hash()
